move tests, rename config module
